refactor: move program log section into separate component
