+++ explorer/src/pages/TransactionDetailsPage.tsx
@@ -27,6 +27,7 @@ import { BigNumber } from "bignumber.js";
 import { BalanceDelta } from "components/common/BalanceDelta";
 import { TokenBalancesCard } from "components/transaction/TokenBalancesCard";
 import { InstructionsSection } from "components/transaction/InstructionsSection";
+import { ProgramLogSection } from "components/transaction/ProgramLogSection";
 
 const AUTO_REFRESH_INTERVAL = 2000;
 const ZERO_CONFIRMATION_BAILOUT = 5;
@@ -393,33 +394,5 @@ function AccountsCard({
         </table>
       </div>
     </div>
-  );
-}
-
-function ProgramLogSection({ signature }: SignatureProps) {
-  const details = useTransactionDetails(signature);
-  const logMessages = details?.data?.transaction?.meta?.logMessages;
-
-  if (!logMessages || logMessages.length < 1) {
-    return null;
-  }
-
-  return (
-    <>
-      <div className="container">
-        <div className="header">
-          <div className="header-body">
-            <h3 className="card-header-title">Program Log</h3>
-          </div>
-        </div>
-      </div>
-      <div className="card">
-        <ul className="log-messages">
-          {logMessages.map((message, key) => (
-            <li key={key}>{message.replace(/^Program log: /, "")}</li>
-          ))}
-        </ul>
-      </div>
-    </>
   );
 }
